requirements.txt
django==3.1
django-python3-ldap==0.11.3
DingtalkChatbot==1.5.1
django-grappelli==2.14.2
django-registration-redux==2.8
django-bootstrap4==2.2.0
djangorestframework==3.11.1
django-filter==2.3.0
Markdown==3.2.2
blinker==1.4
sentry-sdk==0.19.1
django-redis==4.12.1
celery==4.4.7
flower==0.9.5
django-celery-beat==2.0.0
django-oss-storage==1.1.1
django-smart-selects==1.5.9
django-debug-toolbar==3.1.1
Pillow==7.0.0
django-simpleui==2020.9.26
uvicorn==0.12.2
django-prometheus==2.1.0
